refactor: style 분리 코리반영
