+++ src/components/shared/Header/LoginMenu.tsx
@@ -1,4 +1,4 @@
-import { buttonContainer } from "./Header.css";
+import { buttonContainer } from "./index.css";
 
 const LoginMenu = () => {
   const handleProfileClick = () => {
