+++ src/components/shared/Header/Profile.css.ts
@@ -1,0 +1,7 @@
+import { style } from "@vanilla-extract/css";
+
+export const dropdownStyle = style({
+  position: "absolute",
+  top: "6.5rem",
+  right: "6rem",
+});
