+++ src/components/shared/Header/index.tsx
@@ -2,13 +2,13 @@
 
 import { IcnLogo } from "@/assets/svgs";
 import Link from "next/link";
+import LoginMenu from "./LoginMenu";
+import UserMenu from "./UserMenu";
 import {
   headerStyle,
   logoContainer,
   logoStyle
-} from "./Header.css";
-import LoginMenu from "./LoginMenu";
-import UserMenu from "./UserMenu";
+} from "./index.css";
 
 const Header = () => {
   /* TODO: 로그인 api 부착 후 Atom으로 교체 */
